requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
boto3==1.28.55
botocore==1.31.55
bs4==0.0.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
contourpy==1.1.1
crispy-bootstrap5==0.7
cryptography==41.0.4
cycler==0.11.0
defusedxml==0.7.1
dj-database-url==2.1.0
Django==4.2.5
django-allauth==0.57.0
django-cleanup==8.0.0
django-crispy-forms==2.0
django-currencies==0.10.1
django-heroku==0.3.1
django-storages==1.14
docutils==0.20.1
fonttools==4.42.1
forex-python==1.8
gunicorn==21.2.0
idna==3.4
importlib-resources==6.1.0
jmespath==0.7.1
jsonfield==2.1.1
kiwisolver==1.4.5
numpy==1.26.0
oauthlib==3.2.2
packaging==23.1
Pillow==10.0.1
psycopg2==2.9.7
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-decouple==3.8
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.7.0
simplejson==3.19.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==1.26.16
whitenoise==6.5.0
zipp==3.17.0
